+++ src/components/addressSearch/AddressPopup.tsx
@@ -1,13 +1,13 @@
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { addressModalState } from '@/recoil/atom/addressModal';
 
-import * as S from '@/styles/componentsStyles/AddressPopup';
+import * as S from '@/styles/componentsStyles/AddressPopUp.styled';
 import BaseInput from '@/components/ui/BaseInput';
 import { useGenericMutation } from '@/service/mutations/customMutation';
 import { addressesApi } from '@/service/mutations';
+import { PopUpModal } from '../ui';
 const AddressPopup = ({ place }: { place: kakao.maps.services.PlacesSearchResultItem }) => {
-  const modalRef = useRef<HTMLDivElement>(null);
   const [modalOpen, setModalOpen] = useRecoilState(addressModalState);
   const [userInput, setUserInput] = useState<string>('');
   const onRegisterSuccess = () => {
@@ -34,39 +34,22 @@ const AddressPopup = ({ place }: { place: kakao.maps.services.PlacesSearchResult
     registerMutation.mutate(data);
   };
 
-  useEffect(() => {
-    const closeModal = (e: MouseEvent) => {
-      if (modalOpen && modalRef.current && !modalRef.current.contains(e.target as Node)) {
-        // 이벤트가 발생한 노드가 모달 컴포넌트 내부에 존재하지 않는다면 close
-        setModalOpen(false);
-      }
-    };
-
-    // 이벤트 리스너를 document 전체에 붙여줌
-    document.addEventListener('mousedown', closeModal);
-
-    return () => {
-      document.removeEventListener('mousedown', closeModal);
-    };
-  }, [modalOpen, setModalOpen]);
-
+  if (!modalOpen) return null;
   return (
-    <S.Background>
-      <S.Modal ref={modalRef}>
-        <S.Title>주소에 대해 설명해주세요.</S.Title>
-        <S.AddressName>주소 이름</S.AddressName>
-        <BaseInput
-          onChange={(e) => setUserInput(e.target.value)}
-          value={userInput}
-          type="text"
-          placeholder="주소를 지칭할 이름을 지어주세요."
-        />
-        <S.BtnContainer>
-          <S.nextTimeBtn onClick={() => setModalOpen(false)}>다음에 하기</S.nextTimeBtn>
-          <S.changeBtn onClick={onRegisterHandler}>변경하기</S.changeBtn>
-        </S.BtnContainer>
-      </S.Modal>
-    </S.Background>
+    <PopUpModal onClose={() => setModalOpen(false)}>
+      <S.Title>주소에 대해 설명해주세요.</S.Title>
+      <S.Name>주소 이름</S.Name>
+      <BaseInput
+        onChange={(e) => setUserInput(e.target.value)}
+        value={userInput}
+        type="text"
+        placeholder="주소를 지칭할 이름을 지어주세요."
+      />
+      <S.BtnContainer>
+        <S.nextTimeBtn onClick={() => setModalOpen(false)}>다음에 하기</S.nextTimeBtn>
+        <S.changeBtn onClick={onRegisterHandler}>변경하기</S.changeBtn>
+      </S.BtnContainer>
+    </PopUpModal>
   );
 };
