+++ src/components/ui/index.ts
@@ -4,5 +4,6 @@ import Location from "./Location";
 import BaseInput from "./BaseInput";
 import LabelTitle from "./LabelTitle";
 import Button from "./Button";
+import PopUpModal from "./PopUpModal";
 
-export { Title, Divide, Location , BaseInput , LabelTitle, Button};
+export { Title, Divide, Location , BaseInput , LabelTitle, Button , PopUpModal};
